create explain submodule
