+++ mlcolvar/explain/__init__.py
@@ -1,0 +1,9 @@
+__all__ = [
+    "sensitivity_analysis", 
+    "plot_sensitivity",
+    "plot_features_distribution"
+]
+ 
+from .sensitivity import *
+from .utils import *
+# from .lasso import * # lasso requires additional dependencies
