+++ mlcolvar/explain/sensitivity.py
@@ -1,8 +1,10 @@
 import numpy as np
 import torch
+from matplotlib import patches as mpatches
+import matplotlib.pyplot as plt
+import mlcolvar.utils.plot
 
-from mlcolvar.utils.plot import plot_sensitivity
-
+__all__ = [ "sensitivity_analysis", "plot_sensitivity" ]
 
 def sensitivity_analysis(
     model,
@@ -143,6 +145,157 @@ def sensitivity_analysis(
 
     return out
 
+def plot_sensitivity(results, mode="violin", per_class=None, max_features = 100, ax=None):
+    """Plot results of the sensitivity analysis. They can be plotted in three modes:
+    * Violin plot ('violin'), showing the density of per-sample sensitivities besides the mean value
+    * Scatter ('scatter'), plotting the mean and standard deviation of the gradients
+    * Horizontal bar plot ('barh') only displaying the mean of the gradients
+
+    Parameters
+    ----------
+    results : dictionary
+        sensitity results calculated by sensitivity_analysis
+    mode : string, optional
+        ('violin','barh','scatter'), by default 'violin'
+    per_class : bool, optional
+        plot per-class statistics if available, by default plot them if available
+    max_features : int, optional
+        plot at most max_features, by default 100
+    ax : matplotlib axis, optional
+        ax where to plot the results, by default it will be initialized
+
+    See also
+    --------
+    mlcolvar.utils.explain.sensitivity_analysis
+        Perform a sensitivity analysis
+
+    Returns
+    -------
+    ax
+        return the generated matplotlib axis if not passed
+    """
+
+    # retrieve info from results dictionary
+    feature_names = results["feature_names"]
+    n_inputs = len(feature_names)
+    if max_features < n_inputs:
+        print(f'Plotting only the first {max_features} features out of {n_inputs}.')
+        feature_names = feature_names[-max_features:]
+        n_inputs = max_features
+    
+    in_num = np.arange(n_inputs)
+    n_results = len(results["sensitivity"].keys())
+
+    # check whether to plot per-class statistics
+    if per_class is None:
+        per_class = True if n_results > 1 else False
+    else:
+        if not type(per_class) == bool:
+            raise TypeError("per_class should be either `True`, `False` or `None`.")
+        if per_class & (n_results == 1):
+            raise KeyError(
+                "Per class analyis requested but no labels found in the results dictionary. You need to call `sensitivity_analysis` with `per_class`=True. "
+            )
+
+    # initialize ax
+    if ax is None:
+        fig = plt.figure(figsize=(5, 0.25 * n_inputs))
+        ax = fig.add_subplot(111)
+        ax.set_title("Sensitivity Analysis")
+        return_ax = True
+    else:
+        return_ax = False
+
+    # define utils functions
+    def _set_violin_attributes(violin_parts, color, alpha=0.5, label=None, zorder=None):
+        for pc in violin_parts["bodies"]:
+            pc.set_facecolor(color)
+            pc.set_edgecolor(color)
+            pc.set_alpha(alpha)
+            if zorder is not None:
+                pc.set_zorder(zorder)
+        if label is not None:
+            patch_label = (mpatches.Patch(color=color, alpha=alpha), label)
+            return patch_label
+
+    patch_labels = []
+
+    patterns = ["", "", "/", "\\", "|", "-", "+", "x", "o", "O", ".", "*"]
+
+    for i, label in enumerate(results["sensitivity"].keys()):
+        score = results["sensitivity"][label][-max_features:]
+        grad = results["gradients"][label][:,-max_features:]
+
+        color = "fessa0" if "Dataset" in label else f"fessa{7-i}"
+
+        if mode == "barh":
+            alpha = 0.6 if "Dataset" in label else 0.4
+            height = 0.8 if "Dataset" in label else 0.4
+            ax.barh(
+                in_num,
+                score,
+                height=height,
+                color=color,
+                edgecolor="k",
+                hatch=patterns[i],
+                alpha=alpha,
+                label=label,
+            )
+        elif mode == "violin":
+            widths = 0 if (("Dataset" in label) & per_class) else 0.5
+            zorder = 1 if "Dataset" in label else 0
+            showmeans = True if "Dataset" in label else False
+            violin_parts = ax.violinplot(
+                grad,
+                positions=in_num,
+                vert=False,
+                showmeans=showmeans,
+                showextrema=False,
+                widths=widths,
+            )
+            patch_label = _set_violin_attributes(
+                violin_parts, color, alpha=0.5, label=label, zorder=zorder
+            )
+            patch_labels.append(patch_label)
+            if "Dataset" in label:
+                ax.scatter(y=in_num, x=score, c="dimgrey", s=10, zorder=2)
+        elif mode == "scatter":
+            fmt = "D" if "Dataset" in label else "."
+            ax.errorbar(
+                y=in_num,
+                x=score,
+                xerr=grad.std(axis=0),
+                color=color,
+                fmt=fmt,
+                alpha=0.5,
+                label=label,
+            )
+        else:
+            raise NotImplementedError(
+                'plot mode can be only "barh","violin","scatter".'
+            )
+
+        if not per_class:
+            break
+
+    # >> legend
+    ax.set_xlabel("Sensitivity")
+    ax.set_yticks(in_num)
+    ax.set_yticklabels(feature_names, fontsize=9)
+    ax.set_ylabel("Input features")
+
+    if mode == "violin":
+        ax.legend(*zip(*patch_labels), loc="lower right", frameon=False)
+    else:
+        ax.legend(loc="lower right", frameon=False)
+    if np.min(results["sensitivity"]["Dataset"])>=0: 
+        ax.set_xlim(0, None)
+    else:
+        ax.axvline(0,color='grey')
+    ax.set_ylim(-1, in_num[-1] + 1)
+
+    if return_ax:
+        return ax
 
 def test_sensitivity_analysis():
     from mlcolvar.data import DictDataset
